requirements.txt
alembic==1.11.1
autopep8==2.0.2
black==23.3.0
blinker==1.9.0
blis==0.7.11
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
dnspython==2.7.0
email-validator==2.0.0
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.6.0/en_core_web_md-3.6.0-py3-none-any.whl#sha256=1618fc62e7aedc2b778caa3ea6e9c8280beac3b83fe7b93316cbb9e8f3abe7c5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
exceptiongroup==1.2.2
flake8==6.0.0
Flask==2.2.3
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
fonttools==4.57.0
greenlet==3.1.1
gunicorn==20.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.44.0
Mako==1.3.10
marisa-trie==1.2.1
MarkupSafe==3.0.2
matplotlib==3.7.2
mccabe==0.7.0
murmurhash==1.0.12
mypy-extensions==1.0.0
nltk==3.9.1
numba==0.61.2
numpy==1.24.3
nvidia-nccl-cu12==2.26.2.post1
packaging==24.2
pandas==2.0.1
pathlib_abc==0.1.1
pathspec==0.12.1
pathy==0.11.0
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
preshed==3.0.9
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pydantic==1.10.21
pyflakes==3.0.1
pyparsing==3.0.9
pytest==7.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3
regex==2024.11.6
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
shap==0.47.1
six==1.17.0
slicer==0.0.8
smart-open==6.4.0
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.9
srsly==2.5.1
thinc==8.1.12
threadpoolctl==3.6.0
tomli==2.2.1
tqdm==4.67.1
typer==0.9.4
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
wasabi==1.1.3
Werkzeug==2.2.3
WTForms==3.0.1
xgboost==3.0.0
